Failed in unittest on windows.
git-svn-id: http://svn.neotitans.net/lhafile@7 9ad4c1a5-682f-4c36-ab82-9c282c375436

+++ lhafile/lzhlib.c
@@ -46,7 +46,7 @@ static char __author__[] =
 #define inline __inline
 #endif
 
-typedef off_t Py_off_t;
+typedef long long Py_off_t;
 
 #define FILE_BUFFER_SIZE (64*1024)
